requirements.txt
# ==================== Core Analysis Libraries ====================
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter>=1.0.0
notebook>=7.0.0
requests>=2.31.0
openpyxl>=3.1.0
plotly>=5.14.0
scikit-learn>=1.3.0

# ==================== Lottery Prediction System ====================
# Database
supabase==2.10.0

# Web Scraping
beautifulsoup4==4.12.3
lxml==5.1.0

# Telegram Bot
python-telegram-bot==21.0.1

# Utilities
python-dotenv==1.0.1

# ==================== V3 — XGBoost Models ====================
xgboost>=2.0.0
lightgbm>=4.0.0
joblib>=1.3.0
